+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bug fixes:
 * Fix `IO.copy_stream` with a `Tempfile` destination (#3280, @eregon).
 * Fix `Regexp.union` negotiating the wrong result encoding (#3287, @nirvdrum, @simonlevasseur).
 * Fix `Proc#parameters` and return all the numbered parameters lower than the used explicitly ones (@andrykonchin).
+* Fix `IO.select([io], nil, [io])` on macOS, it was hanging due to a bug in macOS `poll(2)` (#3346, @eregon, @andrykonchin).
 
 Compatibility: